#38
- examples/server: пересмотр структуры
- di/token: правки toString
- tokens: переименование типов перед добавлением новых
