+++ examples/server/src/index.ts
@@ -1,8 +1,8 @@
-import { MainApp } from './di/apps/main';
-import { TOKEN } from './di/tokens';
+import { MainApp } from './app';
+import { TOKEN } from './tokens';
 
 MainApp().invoke(
-  [TOKEN.appConfig, TOKEN.Known.logger, TOKEN.httpServer, TOKEN.Known.Metrics.httpApp],
+  [TOKEN.Project.config, TOKEN.Lib.logger, TOKEN.Project.Http.server, TOKEN.Lib.Metrics.httpApp],
   (config, logger, mainServer, metricsServer) => {
     mainServer.listen(config.http.ports.main, () => {
       logger.info(`Server started on port ${config.http.ports.main}`);
