+++ examples/server/src/tokens.ts
@@ -1,16 +1,22 @@
 import { createToken } from '@sima-land/isomorph/di';
 import { KnownToken } from '@sima-land/isomorph/tokens';
-import type { AppConfig, HttpApi } from '../app';
+import type { AppConfig, HttpApi } from './app/types';
 import type { Application, Handler } from 'express';
 
 export const TOKEN = {
   // реэкспорт "известных токенов" для удобного использования
-  Known: KnownToken,
+  Lib: KnownToken,
 
   // токены компонентов нашего приложения
-  appConfig: createToken<AppConfig>('app-config'),
-  httpServer: createToken<Application>('http-server'),
-  usersHandler: createToken<Handler>('users-handler'),
-  postsHandler: createToken<Handler>('posts-handler'),
-  httpApi: createToken<HttpApi>('response/api'),
+  Project: {
+    config: createToken<AppConfig>('config'),
+    Http: {
+      api: createToken<HttpApi>('http/api'),
+      server: createToken<Application>('http/server'),
+      Pages: {
+        users: createToken<Handler>('pages/users'),
+        posts: createToken<Handler>('pages/posts'),
+      },
+    },
+  },
 } as const;
